requirements.txt

streamlit>=1.27.0,<2
google-genai>=0.2.7
numpy>=1.23
PyPDF2>=3.0.0
python-dotenv>=1.0.0